api/requirements.txt
bcrypt==3.1.7
Flask==1.1.2
Flask-Cors==3.0.8
Jinja2==2.11.2
Werkzeug==1.0.1
pymongo==3.10.1
requests==2.23.0
pyjwt==1.7.1
